chore: centralize `NavigationProps`
